requirements.txt
fastapi
streamlit
requests
google-api-python-client
google-auth
google-auth-oauthlib
python-dateutil
transformers
torch
uvicorn